pyproject.toml
[tool.poetry]
name = "nonebot-plugin-soup"
version = "1.1.0"
description = "来一碗心灵鸡汤吧♥"
authors = ["Monarchdos <admin@ayfre.com>"]
readme = "README.md"
homepage = "https://github.com/Monarchdos/nonebot_plugin_soup"
repository = "https://github.com/Monarchdos/nonebot_plugin_soup"
documentation = "https://github.com/Monarchdos/nonebot_plugin_soup#readme"
keywords = ["nonebot", "nonebot2", "ChickenSoup"]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = ">=2.3.3,<3.0.0"
httpx = "^0.27.2"
nonebot-plugin-alconna = "^0.52.2"

[tool.poetry.group.dev.dependencies]
nonebot2 = { extras = ["fastapi"], version = "^2.3.3" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot]
plugins = ["nonebot_plugin_soup"]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
